requirements.txt
google-adk==0.4.0
aiohttp==3.11.18